+++ src/imio/helpers/tests/test_vocabularies.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-
 from imio.helpers.content import get_vocab
 from imio.helpers.testing import IntegrationTestCase
 from plone import api
@@ -27,6 +26,35 @@ class TestVocabularies(IntegrationTestCase):
              u'B_new_user (B_new_user)',
              u'new_user (new_user)',
              u'test_user_1_ (test_user_1_)'])
+        vocab2 = get_vocab(self.portal, "imio.helpers.SimplySortedUsers")
+        self.assertEqual(
+            [term.title for term in vocab2._terms],
+            [u'a_new_user',
+             u'A_new_user',
+             u'b_new_user',
+             u'B_new_user',
+             u'new_user',
+             u'test_user_1_'])
+        member = api.user.get(username="a_new_user")
+        member.setMemberProperties({'fullname': 'anew User'})
+        vocab = get_vocab(self.portal, "imio.helpers.SortedUsers")
+        vocab2 = get_vocab(self.portal, "imio.helpers.SimplySortedUsers")
+        self.assertEqual(
+            [term.title for term in vocab._terms],
+            [u'anew User (a_new_user)',
+             u'A_new_user (A_new_user)',
+             u'b_new_user (b_new_user)',
+             u'B_new_user (B_new_user)',
+             u'new_user (new_user)',
+             u'test_user_1_ (test_user_1_)'])
+        self.assertEqual(
+            [term.title for term in vocab2._terms],
+            [u'anew User',
+             u'A_new_user',
+             u'b_new_user',
+             u'B_new_user',
+             u'new_user',
+             u'test_user_1_'])
 
 
 class EnhancedTermTests(SimpleTermTests):